+++ packages/expo-router/src/typed-routes/index.ts
@@ -4,34 +4,54 @@ import path from 'node:path';
 
 import { getTypedRoutesDeclarationFile } from './generate';
 import { isTypedRoute } from '../matchers';
-import requireContext from '../testing-library/require-context-ponyfill';
+import requireContext, {
+  RequireContextPonyFill,
+} from '../testing-library/require-context-ponyfill';
 
-const ctx = requireContext(process.env.EXPO_ROUTER_APP_ROOT, true, EXPO_ROUTER_CTX_IGNORE);
+export type { RequireContextPonyFill } from '../testing-library/require-context-ponyfill';
 
 /**
  * Generate a Metro watch handler that regenerates the typed routes declaration file
  */
-export function getWatchHandler(outputDir: string) {
+export function getWatchHandler(
+  outputDir: string,
+  ctx = requireContext(process.env.EXPO_ROUTER_APP_ROOT, true, EXPO_ROUTER_CTX_IGNORE),
+  regenerateFn = regenerateDeclarations
+) {
   const routeFiles = new Set(ctx.keys().filter((key) => isTypedRoute(key)));
 
   return async function callback({ filePath, type }: { filePath: string; type: string }) {
+    // Sanity check that we are in an Expo Router project
+    if (!process.env.EXPO_ROUTER_APP_ROOT) return;
+
     let shouldRegenerate = false;
+    let relativePath = path.relative(process.env.EXPO_ROUTER_APP_ROOT, filePath);
+    const isInsideAppRoot = !relativePath.startsWith('../');
+    const basename = path.basename(relativePath);
+
+    if (!isInsideAppRoot) return;
+
+    // require.context paths always start with './' when relative to the root
+    relativePath = `./${relativePath}`;
 
     if (type === 'delete') {
-      ctx.__delete(filePath);
-      if (routeFiles.has(filePath)) {
-        routeFiles.delete(filePath);
+      ctx.__delete(relativePath);
+      if (routeFiles.has(relativePath)) {
+        routeFiles.delete(relativePath);
         shouldRegenerate = true;
       }
     } else if (type === 'add') {
-      ctx.__add(filePath);
-      shouldRegenerate = isTypedRoute(filePath);
+      ctx.__add(relativePath);
+      if (isTypedRoute(basename)) {
+        routeFiles.add(relativePath);
+        shouldRegenerate = true;
+      }
     } else {
-      shouldRegenerate = routeFiles.has(filePath);
+      shouldRegenerate = routeFiles.has(relativePath);
     }
 
     if (shouldRegenerate) {
-      regenerateDeclarations(outputDir);
+      regenerateFn(outputDir, ctx);
     }
   };
 }
@@ -39,7 +59,7 @@ export function getWatchHandler(outputDir: string) {
 /**
  * A throttled function that regenerates the typed routes declaration file
  */
-export const regenerateDeclarations = throttle((outputDir: string) => {
+export const regenerateDeclarations = throttle((outputDir: string, ctx: RequireContextPonyFill) => {
   const file = getTypedRoutesDeclarationFile(ctx);
   if (!file) return;
   fs.writeFileSync(path.resolve(outputDir, './router.d.ts'), file);